Vlv2TbltDevicePkg: Remove double \r
This is a large change, but it shouldn't mess up the git blame. All of
the bad line endings are in the copyright headers.

Note that this commit does not pass the PatchCheck.py presubmit script
due to issues caused by the repeated \r characters.

Command run to fix these files (on Linux):
  find Vlv2TbltDevicePkg -type f -print0 | xargs -0 sed -i "s/\r\r/\r/g"

Cc: Zailiang Sun <zailiang.sun@intel.com>
Cc: Yi Qian <yi.qian@intel.com>
Signed-off-by: Joe Richey <joerichey@google.com>